Move the parse_resource utility to a utils module
